Free CST before test execution on child process
This avoids exposing the internal `cst_exit` function so it can be used on macros. Now, CST frees all of its memory (On the child process only) before running the test, so no leaks are possible on CST's side. This also eliminates the potential risk of introducing memory leaks with new features